refactor: 스터디 룸 UPDATE API Update 제거
