+++ src/main/java/com/jj/swm/domain/studyroom/dto/request/UpdateStudyRoomSettingRequest.java
@@ -1,8 +1,10 @@
 package com.jj.swm.domain.studyroom.dto.request;
 
-import com.jj.swm.domain.studyroom.dto.request.update.option.ModifyStudyRoomOptionInfoRequest;
-import com.jj.swm.domain.studyroom.dto.request.update.reservationType.ModifyStudyRoomReservationTypeRequest;
-import com.jj.swm.domain.studyroom.dto.request.update.type.ModifyStudyRoomTypeInfoRequest;
+import com.jj.swm.domain.studyroom.dto.request.update.ModifyStudyRoomOptionInfoRequest;
+import com.jj.swm.domain.studyroom.dto.request.update.ModifyStudyRoomReservationTypeRequest;
+import com.jj.swm.domain.studyroom.dto.request.update.ModifyStudyRoomTypeInfoRequest;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Positive;
 import lombok.*;
 
 @Getter
@@ -10,6 +12,14 @@ import lombok.*;
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class UpdateStudyRoomSettingRequest {
+
+    @NotNull
+    @Positive
+    private Integer entireMinPricePerHour;
+
+    @NotNull
+    @Positive
+    private Integer entireMaxPricePerHour;
 
     private ModifyStudyRoomOptionInfoRequest optionInfoModification;
