--- src/main/java/com/jj/swm/domain/studyroom/dto/request/update/image/UpdateStudyRoomImageRequest.java
@@ -1,20 +1,0 @@
-package com.jj.swm.domain.studyroom.dto.request.update.image;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
-import lombok.*;
-
-@Getter
-@Builder
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class UpdateStudyRoomImageRequest {
-
-    @NotNull
-    @Positive
-    private Long imageId;
-
-    @NotBlank
-    private String imageUrl;
-}
